♻️ api: explicit query params naming in GET session for supervising route
